- hash returns long.  void * is long.  void * is not int.  bad pjones.

+++ bdevid/python/pybdevid.c
@@ -90,7 +90,7 @@ pybd_bdevid_compare(PyBdevidObject *self, PyBdevidObject *other)
 static long
 pybd_bdevid_hash(PyObject *self)
 {
-    return (int)self;
+    return (long)self;
 }
 
 static PyObject *
